connect to db, refine api

+++ src/main/java/me/bright/user/UserRepository.java
@@ -2,4 +2,4 @@ package me.bright.user;
 
 import org.springframework.data.repository.CrudRepository;
 
-public interface UserRepository extends CrudRepository<User, Long> {}
+public interface UserRepository extends CrudRepository<User, Integer> {}
